Move: Similar Encoders into 1 file
