--- src/lib/kotlin/slatekit-data/src/main/kotlin/slatekit/data/encoders/DoubleEncoder.kt
@@ -1,16 +1,0 @@
-package slatekit.data.encoders
-
-
-import slatekit.common.Record
-import slatekit.data.Consts
-
-open class DoubleEncoder : SqlEncoder<Double> {
-
-    override fun encode(value: Double?): String {
-        return value?.toString() ?: Consts.NULL
-    }
-
-    override fun decode(record: Record, name: String): Double? {
-        return record.getDouble(name)
-    }
-}
