--- src/lib/kotlin/slatekit-data/src/main/kotlin/slatekit/data/encoders/FloatEncoder.kt
@@ -1,16 +1,0 @@
-package slatekit.data.encoders
-
-
-import slatekit.common.Record
-import slatekit.data.Consts
-
-open class FloatEncoder : SqlEncoder<Float> {
-
-    override fun encode(value: Float?): String {
-        return value?.toString() ?: Consts.NULL
-    }
-
-    override fun decode(record: Record, name: String): Float? {
-        return record.getFloat(name)
-    }
-}
