--- src/lib/kotlin/slatekit-data/src/main/kotlin/slatekit/data/encoders/LocalTimeEncoder.kt
@@ -1,18 +1,0 @@
-package slatekit.data.encoders
-
-
-import slatekit.common.Record
-import slatekit.data.Consts
-//import java.time.*
-import org.threeten.bp.*
-
-open class LocalTimeEncoder : SqlEncoder<LocalTime> {
-
-    override fun encode(value: LocalTime?): String {
-        return value?.let { "'" + value.format(Consts.timeFormat) + "'" } ?: Consts.NULL
-    }
-
-    override fun decode(record: Record, name: String): LocalTime? {
-        return record.getLocalTime(name)
-    }
-}
