+++ src/lib/kotlin/slatekit-data/src/main/kotlin/slatekit/data/encoders/UIDEncoders.kt
@@ -2,6 +2,7 @@ package slatekit.data.encoders
 
 
 import slatekit.common.Record
+import slatekit.common.ids.UPID
 import slatekit.data.Consts
 import java.util.*
 
@@ -13,5 +14,17 @@ open class UUIDEncoder : SqlEncoder<UUID> {
 
     override fun decode(record: Record, name: String): UUID? {
         return record.getUUID(name)
+    }
+}
+
+
+open class UPIDEncoder : SqlEncoder<UPID> {
+
+    override fun encode(value: UPID?): String {
+        return value?.let { "'${value.value}'" } ?: Consts.NULL
+    }
+
+    override fun decode(record: Record, name: String): UPID? {
+        return record.getUPID(name)
     }
 }
